refactor soil memory projections
