+++ src/platform/soil/content-projections.ts
@@ -1,4 +1,3 @@
-import * as fsp from "node:fs/promises";
 import * as path from "node:path";
 import { DEFAULT_ROOT, DEFAULT_SEED, DEFAULT_USER } from "../../base/config/identity-loader.js";
 import type {
@@ -8,147 +7,27 @@ import type {
 } from "../../base/types/knowledge.js";
 import type { AgentMemoryEntry, AgentMemoryStore } from "../knowledge/types/agent-memory.js";
 import { AgentMemoryStoreSchema } from "../knowledge/types/agent-memory.js";
-import type { LearnedPattern } from "../knowledge/types/learning.js";
-import type { DreamWorkflowRecord } from "../dream/dream-event-workflows.js";
-import type { SoilMemoryHealthSnapshot } from "./health.js";
-import { computeSoilChecksum } from "./checksum.js";
-import { getDefaultSoilRootDir } from "./config.js";
-import { SoilCompiler } from "./compiler.js";
 import { readTextFileOrNull } from "./io.js";
-import { SoilPageFrontmatterSchema, type SoilPageFrontmatter, type SoilSourceRef } from "./types.js";
+import { SoilPageFrontmatterSchema, type SoilPageFrontmatter } from "./types.js";
+import {
+  baseFrontmatter,
+  nowIso,
+  renderIndexPage,
+  SOIL_COMPILED_MEMORY_SCHEMA_VERSION,
+  soilRootFromBaseDir,
+  sortByDate,
+  sourceHashFromFileOrValue,
+  sourceHashFromText,
+  sourceRefsFromPaths,
+  trimText,
+  type SoilProjectionOptions,
+  writeProjectedPage,
+} from "./projection-support.js";
 
-const SOIL_PROJECTION_VERSION = "soil-v1";
-const SOIL_PAGE_FORMAT_VERSION = "soil-page-v1";
-const SOIL_COMPILED_MEMORY_SCHEMA_VERSION = "soil-compiled-memory-v1";
-
-interface SoilProjectionOptions {
-  baseDir: string;
-  rootDir?: string;
-  clock?: () => Date;
-}
-
-function soilRootFromBaseDir(input: SoilProjectionOptions): string {
-  return input.rootDir ?? getDefaultSoilRootDir(input.baseDir);
-}
-
-function nowIso(clock?: () => Date): string {
-  return (clock?.() ?? new Date()).toISOString();
-}
-
-function sortByDate<T>(values: T[], select: (value: T) => string | undefined): T[] {
-  return [...values].sort((left, right) => {
-    const leftTs = select(left) ?? "";
-    const rightTs = select(right) ?? "";
-    return rightTs.localeCompare(leftTs);
-  });
-}
-
-function trimText(value: string, maxLength = 180): string {
-  const normalized = value.replace(/\s+/g, " ").trim();
-  if (normalized.length <= maxLength) {
-    return normalized;
-  }
-  return `${normalized.slice(0, maxLength - 3)}...`;
-}
-
-function sourceRefsFromPaths(
-  sourceType: SoilSourceRef["source_type"],
-  paths: Array<{ sourcePath: string; sourceHash?: string; reliability?: SoilSourceRef["reliability"] }>
-): SoilSourceRef[] {
-  return paths.map((item) => ({
-    source_type: sourceType,
-    source_path: item.sourcePath,
-    source_hash: item.sourceHash,
-    reliability: item.reliability,
-  }));
-}
-
-function watermarkFromSourceRefs(
-  scope: string,
-  generatedAt: string,
-  sourceRefs: SoilSourceRef[],
-  inputChecksums: Record<string, string>
-): SoilPageFrontmatter["generation_watermark"] {
-  return {
-    scope,
-    source_path: sourceRefs[0]?.source_path,
-    source_paths: sourceRefs.map((ref) => ref.source_path),
-    source_hash: sourceRefs[0]?.source_hash,
-    source_hashes: sourceRefs.map((ref) => ref.source_hash).filter((value): value is string => Boolean(value)),
-    generated_at: generatedAt,
-    projection_version: SOIL_PROJECTION_VERSION,
-    input_commit_ids: [],
-    input_checksums: inputChecksums,
-  };
-}
-
-function baseFrontmatter(input: {
-  soilId: string;
-  title: string;
-  kind: SoilPageFrontmatter["kind"];
-  route: SoilPageFrontmatter["route"];
-  createdAt: string;
-  updatedAt: string;
-  generatedAt: string;
-  sourceRefs: SoilSourceRef[];
-  sourceTruth: SoilPageFrontmatter["source_truth"];
-  renderedFrom: string;
-  summary?: string;
-  goalId?: string;
-  taskId?: string;
-  scheduleId?: string;
-  decisionId?: string;
-  entryId?: string;
-  domain?: string;
-  confidence?: number;
-  priority?: number;
-  inputChecksums?: Record<string, string>;
-}): SoilPageFrontmatter {
-  return SoilPageFrontmatterSchema.parse({
-    soil_id: input.soilId,
-    kind: input.kind,
-    status: "confirmed",
-    title: input.title,
-    route: input.route,
-    source: "compiled",
-    version: "1",
-    created_at: input.createdAt,
-    updated_at: input.updatedAt,
-    generated_at: input.generatedAt,
-    source_refs: input.sourceRefs,
-    generation_watermark: watermarkFromSourceRefs(
-      input.soilId,
-      input.generatedAt,
-      input.sourceRefs,
-      input.inputChecksums ?? {}
-    ),
-    stale: false,
-    manual_overlay: {
-      enabled: false,
-      status: "candidate",
-    },
-    goal_id: input.goalId,
-    task_id: input.taskId,
-    schedule_id: input.scheduleId,
-    decision_id: input.decisionId,
-    entry_id: input.entryId,
-    domain: input.domain,
-    confidence: input.confidence,
-    priority: input.priority,
-    owner: "pulseed",
-    summary: input.summary,
-    source_truth: input.sourceTruth,
-    rendered_from: input.renderedFrom,
-    import_status: "none",
-    approval_status: "none",
-    supersedes: [],
-    page_format_version: SOIL_PAGE_FORMAT_VERSION,
-  });
-}
-
-async function writeProjectedPage(input: SoilProjectionOptions, page: { frontmatter: SoilPageFrontmatter; body: string }): Promise<void> {
-  await SoilCompiler.create({ rootDir: soilRootFromBaseDir(input) }, { clock: input.clock }).write(page);
-}
+export {
+  projectDreamKnowledgeToSoil,
+  projectSoilFeedbackToSoil,
+} from "./compiled-memory-projections.js";
 
 function knowledgeEntrySection(entry: {
   entry_id: string;
@@ -231,117 +110,6 @@ function decisionSection(record: DecisionRecord): string {
     "",
   ];
   return lines.join("\n");
-}
-
-async function sourceHashFromText(content: string | null): Promise<string | undefined> {
-  if (content === null) {
-    return undefined;
-  }
-  return computeSoilChecksum(content);
-}
-
-async function sourceHashFromFileOrValue(sourcePath: string, fallback: unknown): Promise<string> {
-  try {
-    return computeSoilChecksum(await fsp.readFile(sourcePath, "utf-8"));
-  } catch {
-    return computeSoilChecksum(fallback);
-  }
-}
-
-function renderIndexPage(title: string, summary: string, sections: string[]): string {
-  return [`# ${title}`, "", summary, "", ...sections, ""].join("\n");
-}
-
-function learnedPatternSection(pattern: LearnedPattern): string {
-  const lines = [
-    `## ${pattern.pattern_id}`,
-    "",
-    `- Type: ${pattern.type}`,
-    `- Description: ${trimText(pattern.description, 360)}`,
-    `- Confidence: ${pattern.confidence}`,
-    `- Evidence count: ${pattern.evidence_count}`,
-    `- Source goals: ${pattern.source_goal_ids.length > 0 ? pattern.source_goal_ids.map((goalId) => `\`${goalId}\``).join(", ") : "global"}`,
-    `- Applicable domains: ${pattern.applicable_domains.length > 0 ? pattern.applicable_domains.map((domain) => `\`${domain}\``).join(", ") : "none"}`,
-    `- Created: ${pattern.created_at}`,
-    `- Last applied: ${pattern.last_applied_at ?? "none"}`,
-    "",
-  ];
-  return lines.join("\n");
-}
-
-function dreamWorkflowSection(workflow: DreamWorkflowRecord): string {
-  const scopes = workflow.applicability.scopes.length > 0
-    ? workflow.applicability.scopes.map((scope) => `${scope.goal_id ?? "global"}:${scope.task_id ?? "all-tasks"}`)
-    : [
-        ...workflow.applicability.goal_ids.map((goalId) => `${goalId}:all-tasks`),
-        ...workflow.applicability.task_ids.map((taskId) => `global:${taskId}`),
-      ];
-  const lines = [
-    `## ${workflow.workflow_id}`,
-    "",
-    `- Type: ${workflow.type}`,
-    `- Title: ${workflow.title}`,
-    `- Description: ${trimText(workflow.description, 360)}`,
-    `- Confidence: ${workflow.confidence}`,
-    `- Evidence count: ${workflow.evidence_count}`,
-    `- Success / failure: ${workflow.success_count} / ${workflow.failure_count}`,
-    `- Scopes: ${scopes.length > 0 ? scopes.map((scope) => `\`${scope}\``).join(", ") : "global"}`,
-    `- Signals: ${workflow.applicability.signals.length > 0 ? workflow.applicability.signals.map((signal) => `\`${signal}\``).join(", ") : "none"}`,
-    `- Created: ${workflow.created_at}`,
-    `- Updated: ${workflow.updated_at}`,
-    "",
-    "### Preconditions",
-    "",
-    ...(workflow.preconditions.length > 0 ? workflow.preconditions.map((item) => `- ${item}`) : ["- none"]),
-    "",
-    "### Steps",
-    "",
-    ...(workflow.steps.length > 0 ? workflow.steps.map((item) => `- ${item}`) : ["- none"]),
-    "",
-    "### Recovery",
-    "",
-    ...(workflow.recovery_steps.length > 0 ? workflow.recovery_steps.map((item) => `- ${item}`) : ["- none"]),
-    "",
-    "### Evidence",
-    "",
-    ...(workflow.evidence_refs.length > 0 ? workflow.evidence_refs.map((item) => `- ${item}`) : ["- none"]),
-    "",
-  ];
-  return lines.join("\n");
-}
-
-function healthFindingSection(snapshot: SoilMemoryHealthSnapshot): string[] {
-  if (snapshot.findings.length === 0) {
-    return ["No health findings."];
-  }
-  return snapshot.findings.slice(0, 50).flatMap((finding) => [
-    `## ${finding.finding_id}`,
-    "",
-    `- Code: ${finding.code}`,
-    `- Severity: ${finding.severity}`,
-    `- Status: ${finding.status}`,
-    `- Soil ID: ${finding.soil_id ?? "none"}`,
-    `- Record ID: ${finding.record_id ?? "none"}`,
-    `- Route ID: ${finding.route_id ?? "none"}`,
-    `- Source path: ${finding.source_path ?? "none"}`,
-    `- Message: ${trimText(finding.message, 360)}`,
-    "",
-  ]);
-}
-
-function compileMissBucketSection(snapshot: SoilMemoryHealthSnapshot): string[] {
-  if (snapshot.compileMissBuckets.length === 0) {
-    return ["No compile miss buckets."];
-  }
-  return snapshot.compileMissBuckets.slice(0, 50).flatMap((bucket) => [
-    `## ${bucket.key}`,
-    "",
-    `- Count: ${bucket.count}`,
-    `- Reason: ${bucket.reason}`,
-    `- Target path: ${bucket.targetPath ?? "none"}`,
-    `- Route IDs: ${bucket.routeIds.length > 0 ? bucket.routeIds.map((routeId) => `\`${routeId}\``).join(", ") : "none"}`,
-    "",
-  ]);
 }
 
 async function readIdentitySource(baseDir: string, fileName: string, fallback: string): Promise<{ content: string; sourceHash?: string }> {
@@ -530,170 +298,6 @@ export async function projectAgentMemoryToSoil(input: SoilProjectionOptions & { 
   await projectMemoryCategoryPage(input, "procedures", categories.procedures);
   await projectMemoryCategoryPage(input, "lessons", categories.lessons);
   await projectMemoryCategoryPage(input, "patterns", categories.patterns);
-}
-
-export async function projectDreamKnowledgeToSoil(input: SoilProjectionOptions & {
-  learnedPatterns: LearnedPattern[];
-  workflowRecords: DreamWorkflowRecord[];
-}): Promise<void> {
-  const generatedAt = nowIso(input.clock);
-  const learningSourcePath = path.join(input.baseDir, "learning");
-  const workflowSourcePath = path.join(input.baseDir, "dream", "workflows.json");
-  const learningSourceHash = await sourceHashFromFileOrValue(learningSourcePath, input.learnedPatterns);
-  const workflowSourceHash = await sourceHashFromFileOrValue(workflowSourcePath, input.workflowRecords);
-  const learnedPatterns = sortByDate(input.learnedPatterns, (pattern) => pattern.created_at);
-  const workflowRecords = sortByDate(input.workflowRecords, (workflow) => workflow.updated_at);
-
-  const patternsFrontmatter = baseFrontmatter({
-    soilId: "learning/learned-patterns/index",
-    title: "Learned patterns",
-    kind: "knowledge",
-    route: "knowledge",
-    createdAt: learnedPatterns.at(-1)?.created_at ?? generatedAt,
-    updatedAt: learnedPatterns.at(0)?.created_at ?? generatedAt,
-    generatedAt,
-    sourceRefs: sourceRefsFromPaths("runtime_json", [{ sourcePath: learningSourcePath, sourceHash: learningSourceHash, reliability: "high" }]),
-    sourceTruth: "runtime_json",
-    renderedFrom: "dream-consolidator",
-    domain: "learned-patterns",
-    summary: `${learnedPatterns.length} learned patterns`,
-    inputChecksums: { [learningSourcePath]: learningSourceHash },
-  });
-  const patternTypes = [...new Set(learnedPatterns.map((pattern) => pattern.type))].sort();
-  await writeProjectedPage(input, {
-    frontmatter: SoilPageFrontmatterSchema.parse({
-      ...patternsFrontmatter,
-      compiled_memory_schema: SOIL_COMPILED_MEMORY_SCHEMA_VERSION,
-    }),
-    body: [
-      "# Learned patterns",
-      "",
-      `- Patterns: ${learnedPatterns.length}`,
-      `- Types: ${patternTypes.length > 0 ? patternTypes.map((type) => `\`${type}\``).join(", ") : "none"}`,
-      `- Generated: ${generatedAt}`,
-      "",
-      "## Patterns",
-      "",
-      ...(learnedPatterns.length > 0 ? learnedPatterns.map((pattern) => learnedPatternSection(pattern)) : ["No learned patterns."]),
-      "",
-    ].join("\n"),
-  });
-
-  const workflowsFrontmatter = baseFrontmatter({
-    soilId: "dream/workflows/index",
-    title: "Dream workflows",
-    kind: "operations",
-    route: "operations",
-    createdAt: workflowRecords.at(-1)?.created_at ?? generatedAt,
-    updatedAt: workflowRecords.at(0)?.updated_at ?? generatedAt,
-    generatedAt,
-    sourceRefs: sourceRefsFromPaths("runtime_json", [{ sourcePath: workflowSourcePath, sourceHash: workflowSourceHash, reliability: "high" }]),
-    sourceTruth: "runtime_json",
-    renderedFrom: "dream-consolidator",
-    domain: "dream-workflows",
-    summary: `${workflowRecords.length} Dream workflow records`,
-    inputChecksums: { [workflowSourcePath]: workflowSourceHash },
-  });
-  const workflowTypes = [...new Set(workflowRecords.map((workflow) => workflow.type))].sort();
-  await writeProjectedPage(input, {
-    frontmatter: SoilPageFrontmatterSchema.parse({
-      ...workflowsFrontmatter,
-      compiled_memory_schema: SOIL_COMPILED_MEMORY_SCHEMA_VERSION,
-    }),
-    body: [
-      "# Dream workflows",
-      "",
-      `- Workflows: ${workflowRecords.length}`,
-      `- Types: ${workflowTypes.length > 0 ? workflowTypes.map((type) => `\`${type}\``).join(", ") : "none"}`,
-      `- Generated: ${generatedAt}`,
-      "",
-      "## Workflows",
-      "",
-      ...(workflowRecords.length > 0 ? workflowRecords.map((workflow) => dreamWorkflowSection(workflow)) : ["No Dream workflows."]),
-      "",
-    ].join("\n"),
-  });
-}
-
-export async function projectSoilFeedbackToSoil(input: SoilProjectionOptions & {
-  snapshot: SoilMemoryHealthSnapshot;
-}): Promise<void> {
-  const snapshot = input.snapshot;
-  const generatedAt = nowIso(input.clock);
-  const feedbackFrontmatter = baseFrontmatter({
-    soilId: "feedback/context",
-    title: "Soil context feedback",
-    kind: "health",
-    route: "health",
-    createdAt: generatedAt,
-    updatedAt: generatedAt,
-    generatedAt,
-    sourceRefs: [],
-    sourceTruth: "soil",
-    renderedFrom: "soil-feedback",
-    domain: "context-feedback",
-    summary: `${snapshot.compileMissCount} compile misses, ${snapshot.findingCount} findings`,
-  });
-  await writeProjectedPage(input, {
-    frontmatter: SoilPageFrontmatterSchema.parse({
-      ...feedbackFrontmatter,
-      compiled_memory_schema: SOIL_COMPILED_MEMORY_SCHEMA_VERSION,
-    }),
-    body: [
-      "# Soil context feedback",
-      "",
-      `- Generated: ${snapshot.generatedAt}`,
-      `- Compile misses: ${snapshot.compileMissCount}`,
-      `- Finding count: ${snapshot.findingCount}`,
-      "",
-      "# Compile Miss Buckets",
-      "",
-      ...compileMissBucketSection(snapshot),
-      "# Findings",
-      "",
-      ...healthFindingSection(snapshot),
-      "",
-    ].join("\n"),
-  });
-
-  const healthFrontmatter = baseFrontmatter({
-    soilId: "health",
-    title: "Soil health",
-    kind: "health",
-    route: "health",
-    createdAt: generatedAt,
-    updatedAt: generatedAt,
-    generatedAt,
-    sourceRefs: [],
-    sourceTruth: "soil",
-    renderedFrom: "soil-feedback",
-    domain: "soil-health",
-    summary: `${snapshot.errorCount} errors, ${snapshot.warningCount} warnings`,
-  });
-  await writeProjectedPage(input, {
-    frontmatter: SoilPageFrontmatterSchema.parse({
-      ...healthFrontmatter,
-      compiled_memory_schema: SOIL_COMPILED_MEMORY_SCHEMA_VERSION,
-    }),
-    body: [
-      "# Soil health",
-      "",
-      "Health checks for the compiled Soil memory layer.",
-      "",
-      `- Generated: ${snapshot.generatedAt}`,
-      `- Root: ${snapshot.rootDir}`,
-      `- Pages: ${snapshot.totalPages}`,
-      `- Findings: ${snapshot.findingCount}`,
-      `- Errors: ${snapshot.errorCount}`,
-      `- Warnings: ${snapshot.warningCount}`,
-      `- Compile misses: ${snapshot.compileMissCount}`,
-      "",
-      "## Findings",
-      "",
-      ...healthFindingSection(snapshot),
-      "",
-    ].join("\n"),
-  });
 }
 
 function decisionSourcePath(baseDir: string, record: DecisionRecord): string {
